moving translation code to separate
