+++ WordPressService/index.js
@@ -10,6 +10,10 @@ const {
     gitHubFileGet,
     gitHubFileGetBlob
 } = require('./gitHub');
+const {
+    addTranslationPings,
+    postTranslations
+} = require('./avantPage');
 
 const { JSDOM } = require('jsdom');
 const sha1 = require('sha1');
@@ -29,28 +33,13 @@ const shaupdate = (file, wp_sha, github_sha) => {
 
 let pinghistory = []; //Used to log updates
 
-//const masterbranch='synctest3', stagingbranch='synctest3_staging', postTranslationUpdates = false, branchprefix = 'synctest3_deploy_';
-const masterbranch='master', stagingbranch='staging', postTranslationUpdates = true, branchprefix = 'wpservice_deploy_';
-const autoApproveTranslationPrs = true;
+const masterbranch='synctest3', stagingbranch='synctest3_staging', postTranslationUpdates = true;
+//const masterbranch='master', stagingbranch='staging', postTranslationUpdates = true;
 const mergetargets = [masterbranch,stagingbranch];
 const appName = 'WordPressService';
-const githubTranslationPingsPath = `pages/translations/pings`;
-const githubTranslationContentPath = `pages/translations/content`;
-const githubTranslationFlatPath = `pages/translated-posts`;
 const githubSyncFolder = 'pages/wordpress-posts'; //no slash at the end
 const wordPressUrl = 'https://as-go-covid19-d-001.azurewebsites.net';
 const wordPressApiUrl = `${wordPressUrl}/wp-json/wp/v2/`;
-const translationUpdateEndpointUrl = 'https://workflow.avant.tools/subscribers/xtm';
-const translationDownloadUrl = `https://storage.googleapis.com/covid19-ca-files-avantpage/`;
-const translatedLanguages = [
-    {code:'ar_AA',tag:'lang-ar',slugpostfix:'ar'},
-    {code:'es_US',tag:'lang-es',slugpostfix:'es'},
-    {code:'ko_KR',tag:'lang-ko',slugpostfix:'ko'},
-    {code:'tl_PH',tag:'lang-tl',slugpostfix:'tl'},
-    {code:'vi_VN',tag:'lang-vi',slugpostfix:'vi'},
-    {code:'zh_TW',tag:'lang-zh-Hant',slugpostfix:'zh-hant'},
-    {code:'zh_CN',tag:'lang-zh-Hans',slugpostfix:'zh-hans'}
-];
 const defaultTags = [];
 const ignoreFiles = []; //No longer needed since manual-content folder used.
 const tag_ignore = 'do-not-deploy';
@@ -62,7 +51,6 @@ const tag_nocrawl = 'do-not-crawl';
 const tag_langprefix = 'lang-';
 const tag_langdefault = 'en';
 const tag_nomaster = 'staging-only';
-const TranslationPrLabels = ['Translated Content'];
 
 module.exports = async function (context, req) {
 
@@ -81,7 +69,7 @@ if(req.method==='GET') {
 
 //Logging data
 const started = getPacificTimeNow();
-let add_count = 0, update_count = 0, delete_count = 0, binary_match_count = 0, sha_match_count = 0, ignore_count = 0, staging_only_count = 0, translation_pings_count = 0, translation_files_count = 0;
+let add_count = 0, update_count = 0, delete_count = 0, binary_match_count = 0, sha_match_count = 0, ignore_count = 0, staging_only_count = 0;
 
 //Translation Update
 const translationUpdatePayload = [];
@@ -108,7 +96,7 @@ const translationUpdateAddPost = (Post, download_path) => {
 }
 
 const branchCreate_WithName = async (filename,mergetarget) => {
-    const branch = mergetarget + branchprefix + filename;
+    const branch = mergetarget + '_wpservice_deploy_' + filename;
     await gitHubBranchCreate(branch,mergetarget);
     return branch;
 }
@@ -263,154 +251,10 @@ for(const mergetarget of mergetargets) {
     }
 }
 
-
-
-
-
-const getTranslatedPageData = html => {
-    //clean up any input issues
-
-    //remove arabic reverse (RTL override) if it is at the beginning
-    while (html.charCodeAt(0)===8294) html=html.substring(1);
-
-    return html.trimLeft();
-}
-
-//Add translation pings
-const addTranslationPings = async () => {
-    if(!req.body||req.headers['content-type']!=='application/json') return;
-
-    for(const mergetarget of mergetargets) {
-
-        const pingJSON = typeof req.body === "string" ? JSON.parse(req.body) : req.body;
-
-        const files_id = pingJSON.files_id;
-        const newFileName = `ping-${files_id}-${new Date().getTime()}.json`;
-        const newFilePath = `${githubTranslationPingsPath}/${newFileName}`;
-        const branch = await branchCreate_WithName(`ping-${files_id}`,mergetarget);
-        const content =  Buffer.from(JSON.stringify(pingJSON,null,2)).toString('base64');
-        await gitHubFileAdd(
-            content,
-            newFilePath,
-            gitHubMessage('Add translation ping',newFileName),
-            branch
-            )
-            .then(() => {console.log(`Add translation ping Success: ${newFileName}`);});
-        translation_pings_count++;
-
-        const translated_on = new Date(pingJSON.translated_on*1000);
-        const posts = pingJSON.posts.map(x=>Number(x));
-
-        if(!files_id||!translated_on||!posts) return;
-
-        const sourceFiles = [];
-        for(const post_id of posts) {
-            const manifestrecord = manifest.posts.find(p=>p.id===post_id);
-
-            if(manifestrecord) {
-                const slug = manifestrecord.slug;
-                sourceFiles.push(slug);
-
-                for(const langRow of translatedLanguages) {
-                    const newslug = `${slug}-${langRow.slugpostfix}`;
-
-                    const downloadContentName = `${slug}-${langRow.code}.html`;
-                    const downloadFilePath = `${files_id}/${post_id}/${downloadContentName}`;
-                    const downloadURL = `${translationDownloadUrl}${downloadFilePath}`;
-
-                    const file = await fetch(downloadURL);
-                    
-                    if(file.status!==200) {
-                        //Can't find the lang file
-                        console.log(`FETCH FILE ERROR ${file.status} - ${downloadFilePath}`);
-                    } else {
-                        console.log(`processing...${downloadFilePath}`);
-                        translation_files_count++;
-
-                        const html = getTranslatedPageData(await file.text());
-
-                        let contentString = '';
-                        if(manifestrecord.isTableData)
-                            contentString = JsonFromHtmlTables(html);
-                        else if(manifestrecord.isFragment)
-                            contentString = html;
-                        else {
-                            //replace the 'translate' tag with the correct lang tag
-
-                            contentString = html.replace(/\"translate\"/,`\"translate\"\,\"${langRow.tag}\"`);
-                        }
-                        const content = Buffer.from(contentString).toString('base64');
-
-                        const newContentName = `${newslug}.${manifestrecord.isTableData ? 'json' : 'html'}`;
-                        const newContentPath = `${githubTranslationContentPath}/${files_id}/${post_id}/${newContentName}`;
-
-                        const existingContent = await gitHubFileGet(newContentPath,branch);
-                        if(existingContent.sha) {
-                            const json = existingContent;
-                            await gitHubFileUpdate(
-                                content,
-                                json.url,
-                                json.sha,
-                                gitHubMessage('Update translation content',newContentName),
-                                branch
-                            )
-                            .then(() => {console.log(`Update translation content Success: ${newContentName}`);});
-                        } else {
-                            await gitHubFileAdd(
-                                content,
-                                newContentPath,
-                                gitHubMessage('Add translation content',newContentName),
-                                branch
-                            )
-                            .then(() => {console.log(`Add translation content Success: ${newContentName}`);});
-                        }
-
-                        const newURL = `${githubTranslationFlatPath}/${newContentName}`;
-
-                        const existingFileResponse = await gitHubFileGet(newURL,branch);
-
-                        if(existingFileResponse.sha) {
-                            //update
-                            const json = existingFileResponse;
-
-                            await gitHubFileUpdate(
-                                content,
-                                json.url,
-                                json.sha,
-                                gitHubMessage('Update translation',newContentName) + `\nSource : ${downloadURL}`,
-                                branch
-                            );
-
-                            console.log(`UPDATE Success: ${newContentName}`);
-                        } else {
-                            //new
-                            await gitHubFileAdd(
-                                content,
-                                `${githubTranslationFlatPath}/${newContentName}`,
-                                gitHubMessage('Add translation',newContentName) + `\nSource : ${downloadURL}`,
-                                branch
-                            );
-                            console.log(`ADD Success: ${newContentName}`);
-                        }
-                    }
-                }
-            }
-        }
-
-        await gitHubBranchMerge(
-            branch,
-            mergetarget,
-            mergetarget===masterbranch,
-            `Translation - ${sourceFiles.join(`, `)}`,
-            TranslationPrLabels,
-            autoApproveTranslationPrs
-            );
-    } //for
-}
-await addTranslationPings();
+await addTranslationPings(manifest,mergetargets,req);
 
 //Add to log
-const total_changes = add_count+update_count+delete_count+translation_pings_count+translation_files_count;
+const total_changes = add_count+update_count+delete_count;
 const log = {
     sourcebranch: masterbranch,
     runtime: `${started} to ${getPacificTimeNow()}`
@@ -422,8 +266,6 @@ if(sha_match_count>0) log.sha_match_count = sha_match_count;
 if(add_count>0) log.add_count = add_count;
 if(update_count>0) log.update_count = update_count;
 if(delete_count>0) log.delete_count = delete_count;
-if(translation_pings_count>0) log.translation_pings_count = translation_pings_count;
-if(translation_files_count>0) log.translation_files_count = translation_files_count;
 if(ignore_count>0) log.ignore_count = ignore_count;
 if(staging_only_count>0) log.staging_only_count = staging_only_count;
 if(total_changes>0) log.total_changes = total_changes;
@@ -440,13 +282,8 @@ pinghistory.unshift(log);
         }
     };
 
-    if(postTranslationUpdates&&translationUpdatePayload.length>0) {
-        const postTranslationOptions = {
-            method: 'POST',
-            body: JSON.stringify({posts:translationUpdatePayload})
-        };
-        await fetch(translationUpdateEndpointUrl, postTranslationOptions)
-            .then(() => {console.log(`Translation Update POST Success`);})
+    if(postTranslationUpdates&&translationUpdatePayload.length) {
+        await postTranslations(translationUpdatePayload);
     }
 
     console.log('done.');
